Move Secret into secrets package
(cherry picked from commit 8f0343e70a630e27bac82aa57cb2a37867d0a8fb)
