+++ reporting/report_test.go
@@ -3,6 +3,8 @@ package reporting
 import (
 	"reflect"
 	"testing"
+
+	"github.com/checkmarx/2ms/secrets"
 )
 
 func TestAddSecretToFile(t *testing.T) {
@@ -23,9 +25,9 @@ JPcHeO7M6FohKgcEHX84koQDN98J/L7pFlSoU7WOl6f8BKavIdeSTPS9qQYWdQuT
 4Xgur9w/aLZrLM3DSatR+kL+cVTyDTtgCt9Dc8k48Q==
 -----END RSA PRIVATE KEY-----`)
 
-	results := map[string][]*Secret{}
+	results := map[string][]*secrets.Secret{}
 	report := Report{len(results), 1, results}
-	secret := &Secret{Source: "bla", StartLine: 0, StartColumn: 0, EndLine: 0, EndColumn: 0, Value: secretValue}
+	secret := &secrets.Secret{Source: "bla", StartLine: 0, StartColumn: 0, EndLine: 0, EndColumn: 0, Value: secretValue}
 	source := "directory\\rawStringAsFile.txt"
 
 	report.Results[source] = append(report.Results[source], secret)
