+++ secrets/engine_test.go
@@ -6,7 +6,6 @@ import (
 	"testing"
 
 	"github.com/checkmarx/2ms/plugins"
-	"github.com/checkmarx/2ms/reporting"
 	"github.com/checkmarx/2ms/secrets/rules"
 )
 
@@ -123,7 +122,7 @@ func TestSecrets(t *testing.T) {
 		}
 		t.Run(name, func(t *testing.T) {
 			fmt.Printf("Start test %s", name)
-			secretsChan := make(chan *reporting.Secret, 1)
+			secretsChan := make(chan *Secret, 1)
 			wg := &sync.WaitGroup{}
 			wg.Add(1)
 			detector.Detect(plugins.Item{Content: secret.Content}, secretsChan, wg, nil)
